[CWS] Try to fix sk_protocol bitfield access

+++ pkg/security/ebpf/c/include/structs/syscalls.h
@@ -266,7 +266,7 @@ struct syscall_cache_t {
 
         struct {
             short socket_type;
-            u16 socket_protocol;
+            unsigned int socket_protocol;
             u16 socket_family;
             unsigned short filter_len;
             u16 filter_size_to_send;
